+++ Library/OcMainLib/OpenCoreUefi.c
@@ -816,6 +816,7 @@ OcLoadUefiSupport (
   IN UINT8               *Signature
   )
 {
+  EFI_STATUS            Status;
   EFI_HANDLE            *DriversToConnect;
   EFI_EVENT             Event;
   BOOLEAN               AvxEnabled;
@@ -839,6 +840,11 @@ OcLoadUefiSupport (
 
   if (Config->Uefi.Quirks.IgnoreInvalidFlexRatio) {
     OcCpuCorrectFlexRatio (CpuInfo);
+  }
+
+  if (Config->Uefi.Quirks.EnableVmx) {
+    Status = OcCpuEnableVmx ();
+    DEBUG ((EFI_ERROR (Status) ? DEBUG_WARN : DEBUG_INFO, "OC: Enable VMX - %r\n", Status));
   }
 
   if (Config->Uefi.Quirks.TscSyncTimeout > 0) {